Notify_py
This project has no review commit to have a readme and a review for this project and please refer the code of the c version for this project. and have a review which differentiates the c programing with advanced programming techniques which can help in differentiating things ---- ITS INPRTANT TO HAVE SOME README THAT STATES the purpose of the program "differentiating b/w the function based and the class based implementation of a project ." finally point out the differences we have encountered b/w both the languages.